requirements.txt
python-dotenv
llama_parse
llama_index
streamlit